Implement new extension API to support Alloca with addresspace. (#73)
* This allow generators using address spaces to directly support allocas with a different address space ID. The actual address won't change but it is usable to indicate that
an allocation results in a value valid for an address space. This is normally used for runtimes that support some sort of GC/managed reference vs a native pointer. It is useful to distinguish the intended use and that the address space is part of the type of the pointer (even if it is OPAQUE).
* Cleaned up some headers
    - Minor re-formatting
    - Protected wrappers for `<cstdint>` and `<climits>` as those are not formally available for ISO/C compilation.
        - While it is increasingly rare to use a pure C compilation it is plausible and should still be supported.